+++ packages/client/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @phantom/client
 
+## 1.0.0-beta.0
+
+### Major Changes
+
+- New SDKs chain specific interface, prepare for initial beta release
+
+### Patch Changes
+
+- ef68eaf: Key rotation
+- Updated dependencies [ef68eaf]
+- Updated dependencies
+  - @phantom/sdk-types@1.0.0-beta.0
+  - @phantom/constants@1.0.0-beta.0
+  - @phantom/api-key-stamper@1.0.0-beta.0
+  - @phantom/base64url@1.0.0-beta.0
+  - @phantom/crypto@1.0.0-beta.0
+
+## 0.1.11
+
+### Patch Changes
+
+- b20bc21: Support derivation index
+
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @phantom/sdk-types@0.1.5
+
 ## 0.1.9
 
 ### Patch Changes